refactor: move simctl device to ios-driver
